// src/lib/service/api-client/model/getReportStatsProviderType.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * 핀플로우 프로모션: Admin API
 * 핀플로우 프로모션: Admin API 문서
 * OpenAPI spec version: 1.0
 */

export type GetReportStatsProviderType =
  (typeof GetReportStatsProviderType)[keyof typeof GetReportStatsProviderType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetReportStatsProviderType = {
  MEZZO: 'MEZZO',
  ADPOPCORN: 'ADPOPCORN',
  DAWIN: 'DAWIN',
  GOOGLE: 'GOOGLE',
} as const;